% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Permission.R, R/r6_Task.R
\name{add_permission}
\alias{add_permission}
\title{add permission}
\usage{
add_permission(task)

add_permission(task)
}
\arguments{
\item{task}{a Task R6 class}

\item{name}{the name of the permission}

\item{permission}{a Permission R6 class}
}
\description{
add permission

add task
}
