% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SchemaRedirect.r
\name{censor_function_factory_nothing}
\alias{censor_function_factory_nothing}
\title{censor 0-4 function factory}
\usage{
censor_function_factory_nothing(
  column_name_to_be_censored,
  column_name_value = column_name_to_be_censored
)
}
\arguments{
\item{column_name_to_be_censored}{Name of the column to be censored}

\item{column_name_value}{Name of the column whose value is determining if something should be censored}
}
\description{
censor 0-4 function factory
}
