% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\name{copy_into_new_table_where}
\alias{copy_into_new_table_where}
\title{copy_into_new_table_where
Keeps the rows where the condition is met}
\usage{
copy_into_new_table_where(conn = NULL, table_from, table_to, condition = "1=1")
}
\arguments{
\item{conn}{A db connection}

\item{table_from}{Table name}

\item{table_to}{Table name}

\item{condition}{A string SQL condition}
}
\description{
copy_into_new_table_where
Keeps the rows where the condition is met
}
