% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create}
\alias{create}
\alias{sc_from_data_frame}
\alias{sc_from_list}
\alias{reduce_sc}
\title{Creating simplicial complexes}
\usage{
sc_from_data_frame(d, vertices = NULL, reduce = FALSE)

sc_from_list(l, vertices = NULL, reduce = FALSE)

reduce_sc(sc)
}
\arguments{
\item{d}{A data frame of vertex-face membership with columns \code{vertex} and
\code{simplex}.}

\item{vertices}{A character or integer vector of vertex IDs, possibly
including more than those included in \code{d} or \code{l}.}

\item{reduce}{Whether to iteratively remove simplices that are faces of other
simplices.}

\item{l}{A list of simplices consisting of integer member vertex IDs.}

\item{sc}{A simplicial complex (an \link[igraph:igraph]{igraph::igraph} object prefixed with
class 'sc').}
}
\value{
A simplicial complex encoded as an \link[igraph:igraph]{igraph::igraph} object prefixed
with class 'sc'.
}
\description{
These functions create 'sc' objects from various input sources.
}
\examples{
# a face list with some redundancies
x <- list(a = 1:4, b = 3:5, c = c(3,6), d = 7, e = c(3,6), f = c(3:5, 7))
# creating a simplicial complex removes duplicates
(sc <- sc_from_list(x))
plot(sc)
# further reduction can be done by testing for subset relations among the faces
(rsc <- reduce_sc(sc))
plot(rsc)
}
