% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingmethods_pt1.R
\name{delta.method}
\alias{delta.method}
\title{Delta Method for Ratio Estimation}
\usage{
delta.method(num, dem)
}
\arguments{
\item{num}{Numeric vector of numerator values}

\item{dem}{Numeric vector of denominator values}
}
\value{
The estimated standard error of the ratio
}
\description{
Computes the standard error of a ratio using the delta method.
}
\details{
This function implements the delta method to estimate the standard error
of a ratio of two random variables. It assumes the ratio is of the form num/dem.
}
\examples{
set.seed(123)
numerator <- rnorm(100, mean = 5, sd = 1)
denominator <- rnorm(100, mean = 2, sd = 0.5)
delta.method(numerator, denominator)

}
