% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_config_last_updated.r
\name{get_config_datetime}
\alias{get_config_datetime}
\title{get_config_datetime
Gets the config_datetime db table}
\usage{
get_config_datetime(type = NULL, tag = NULL)
}
\arguments{
\item{type}{a}

\item{tag}{a}
}
\description{
get_config_datetime
Gets the config_datetime db table
}
