% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingmethods_pt1.R
\name{inv_cdf}
\alias{inv_cdf}
\title{Inverse CDF Sampling}
\usage{
inv_cdf(inv, n = 1e+05)
}
\arguments{
\item{inv}{A function representing the inverse CDF of the target distribution.}

\item{n}{The number of samples to generate.}
}
\value{
A numeric vector of sampled values.
}
\description{
This function generates random samples from a given continuous probability distribution
using the inverse cumulative distribution function (CDF) method.
}
\examples{
# Define the inverse CDF for an Exponential Distribution
inv_cdf_exp <- function(u) { -log(1 - u) }

# Generate 1000 samples using Inverse CDF Sampling
samples <- inv_cdf(inv_cdf_exp, n = 1000)

# Plot the comparison between the exponential PDF and sampled values
plot_sampling(samples, function(x) dexp(x))  # Compare with exponential PDF

}
