% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingmethods_pt1.R
\name{inverse_cdf_sampling}
\alias{inverse_cdf_sampling}
\title{Inverse CDF Sampling}
\usage{
inverse_cdf_sampling(n, inv_cdf)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{inv_cdf}{A function representing the inverse CDF of the target distribution.}
}
\value{
A numeric vector of sampled values.
}
\description{
This function generates random samples from a given continuous probability distribution
using the inverse cumulative distribution function (CDF) method.
}
\examples{
inv_cdf_exp <- function(u) { -log(1 - u) }  # Exponential(1) inverse CDF
samples <- inverse_cdf_sampling(1000, inv_cdf_exp)
plot_sampling(samples, function(x) dexp(x))  # Compare with exponential PDF
}
