% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_folders.R
\name{path}
\alias{path}
\title{This function gets the right folder for results}
\usage{
path(type = "output", ...)
}
\arguments{
\item{type}{input, output}

\item{...}{Second level and beyond}
}
\description{
This function gets the right folder for results
}
