% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.r
\name{sc-plot}
\alias{sc-plot}
\alias{plot.sc}
\title{Plotting simplicial complexes}
\usage{
\method{plot}{sc}(x, simplex.colors = "viridis", simplex.alpha = 0.5,
  layout = layout_with_fr, ...)
}
\arguments{
\item{x}{Bipartite \link[igraph:igraph]{igraph::igraph} object, representing a simplicial
complex.}

\item{simplex.colors}{Character; the color palette from which to assign
colors to the simplices (by dimension). Matched to one of the \strong{viridis}
palettes (\code{"viridis"}, \code{"magma"}, \code{"plasma"}, \code{"inferno"}) or
\code{"gray.colors"}.}

\item{simplex.alpha}{Numeric value between \code{0} and \code{1} indicating the
transparency of the simplices.}

\item{layout}{A \link[igraph:layout]{igraph::layout} function, or the result of such a function
on the projection of the simplicial complex \code{x}.}

\item{...}{Additional parameters passed to other \code{\link[graphics:plot]{graphics::plot()}} methods.}
}
\description{
These functions use \link[igraph:igraph.plotting]{igraph::igraph.plotting} functionality to
plot a simplicial complex, i.e. a bipartite \link[igraph:igraph]{igraph::igraph} object
prefixed with class 'sc'.
}
\examples{
# example simplicial complex from Wikimedia Commons
# https://commons.wikimedia.org/wiki/File:Simplicial_complex_example.png
wiki_sc <- sc_from_list(list(
  a = c(1,3,4),
  b = 1:2,
  c = 5,
  d = 6:9,
  e = c(6, 10:11),
  f = 10:12,
  g = c(11, 13:14),
  h = 14:15,
  i = c(14, 16),
  j = 15:16,
  k = c(13, 17),
  l = 17:18
))
print(wiki_sc)
plot(wiki_sc, vertex.label = NA)
}
