% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.r
\name{sc-print}
\alias{sc-print}
\alias{print.sc}
\title{Printing simplicial complexes}
\usage{
\method{print}{sc}(x, ...)
}
\arguments{
\item{x}{Bipartite \link[igraph:igraph]{igraph::igraph} object, representing a simplicial
complex.}

\item{...}{Additional parameters passed to \code{\link[tibble:trunc_mat]{tibble::trunc_mat()}}.}
}
\description{
These functions adapt \code{\link[tibble:format.tbl]{tibble::format.tbl()}},
\code{\link[tibble:print.tbl]{tibble::print.tbl()}}, and the unexported function
\code{node_tibble()} from \strong{\link[tidygraph:tidygraph]{tidygraph::tidygraph}} to conveniently display data separately on the
observation and cluster nodes of the bipartite \link[igraph:igraph]{igraph::igraph} object that
encodes a simplicial complex.
}
\examples{
# example simplicial complex from Wikimedia Commons
# https://commons.wikimedia.org/wiki/File:Simplicial_complex_example.png
wiki_sc <- sc_from_list(list(
  a = c(1,3,4),
  b = 1:2,
  c = 5,
  d = 6:9,
  e = c(6, 10:11),
  f = 10:12,
  g = c(11, 13:14),
  h = 14:15,
  i = c(14, 16),
  j = 15:16,
  k = c(13, 17),
  l = 17:18
))
# inspection and visualization
print(wiki_sc)
plot(wiki_sc, vertex.label = NA)
# vertices and simplices
vertices(wiki_sc)
simplices(wiki_sc)
vertices(wiki_sc, simplices = letters[1:3])
simplices(wiki_sc, vertices = 1:5)
}
