% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_test.r
\name{sc_test_plan_argset}
\alias{sc_test_plan_argset}
\title{sc_test (plan/argset)
This function can be deleted if you are not using "plan_argset_fn_name"
inside sc::task_from_config_v3}
\usage{
sc_test_plan_argset(argset, schema)
}
\description{
sc_test (plan/argset)
This function can be deleted if you are not using "plan_argset_fn_name"
inside sc::task_from_config_v3
}
