% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{Schema_v8}
\alias{Schema_v8}
\title{schema class description}
\description{
schema class description

schema class description
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conn}}{db connection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Schema_v8$new()}}
\item \href{#method-connect}{\code{Schema_v8$connect()}}
\item \href{#method-disconnect}{\code{Schema_v8$disconnect()}}
\item \href{#method-create_table}{\code{Schema_v8$create_table()}}
\item \href{#method-drop_table}{\code{Schema_v8$drop_table()}}
\item \href{#method-insert_data}{\code{Schema_v8$insert_data()}}
\item \href{#method-upsert_data}{\code{Schema_v8$upsert_data()}}
\item \href{#method-drop_all_rows}{\code{Schema_v8$drop_all_rows()}}
\item \href{#method-drop_rows_where}{\code{Schema_v8$drop_rows_where()}}
\item \href{#method-keep_rows_where}{\code{Schema_v8$keep_rows_where()}}
\item \href{#method-drop_all_rows_and_then_upsert_data}{\code{Schema_v8$drop_all_rows_and_then_upsert_data()}}
\item \href{#method-tbl}{\code{Schema_v8$tbl()}}
\item \href{#method-list_indexes_db}{\code{Schema_v8$list_indexes_db()}}
\item \href{#method-add_indexes}{\code{Schema_v8$add_indexes()}}
\item \href{#method-drop_indexes}{\code{Schema_v8$drop_indexes()}}
\item \href{#method-clone}{\code{Schema_v8$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Schema_v3 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$new(
  conn = NULL,
  db_config = NULL,
  table_name,
  field_types,
  keys,
  indexes = NULL,
  validator_field_types = validator_field_types_blank,
  validator_field_contents = validator_field_contents_blank,
  info = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{Name.}

\item{\code{db_config}}{Hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Schema_v3` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to a db
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$connect(db_config = self$db_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_config}}{db_config}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from a db
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_table"></a>}}
\if{latex}{\out{\hypertarget{method-create_table}{}}}
\subsection{Method \code{create_table()}}{
Create db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$create_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_table"></a>}}
\if{latex}{\out{\hypertarget{method-drop_table}{}}}
\subsection{Method \code{drop_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$drop_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert_data"></a>}}
\if{latex}{\out{\hypertarget{method-insert_data}{}}}
\subsection{Method \code{insert_data()}}{
Inserts data into db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$insert_data(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-upsert_data}{}}}
\subsection{Method \code{upsert_data()}}{
Upserts data into db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_all_rows"></a>}}
\if{latex}{\out{\hypertarget{method-drop_all_rows}{}}}
\subsection{Method \code{drop_all_rows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$drop_all_rows()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-drop_rows_where}{}}}
\subsection{Method \code{drop_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$drop_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-keep_rows_where}{}}}
\subsection{Method \code{keep_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$keep_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_all_rows_and_then_upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-drop_all_rows_and_then_upsert_data}{}}}
\subsection{Method \code{drop_all_rows_and_then_upsert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$drop_all_rows_and_then_upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-tbl}{}}}
\subsection{Method \code{tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$tbl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_indexes_db"></a>}}
\if{latex}{\out{\hypertarget{method-list_indexes_db}{}}}
\subsection{Method \code{list_indexes_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$list_indexes_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_indexes"></a>}}
\if{latex}{\out{\hypertarget{method-add_indexes}{}}}
\subsection{Method \code{add_indexes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$add_indexes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_indexes"></a>}}
\if{latex}{\out{\hypertarget{method-drop_indexes}{}}}
\subsection{Method \code{drop_indexes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$drop_indexes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema_v8$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
