% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_from_config_v3.R
\name{task_from_config_v3}
\alias{task_from_config_v3}
\title{Creating a task from config (v3)}
\usage{
task_from_config_v3(
  name,
  cores = 1,
  for_each_plan,
  for_each_argset = NULL,
  universal_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action_fn_name,
  data_selector_fn_name = NULL,
  schema = NULL,
  info = NULL
)
}
\arguments{
\item{name}{Name of the task}

\item{cores}{Number of CPU cores}

\item{for_each_plan}{A list, where each unit corresponds to one data extraction. Generally recommended to use \code{plnr::expand_list}.}

\item{for_each_argset}{A list, where each unit corresponds to one analysis within a plan (data extraction). Generally recommended to use \code{plnr::expand_list}.}

\item{universal_argset}{A list, where these argsets are applied to all analyses univerally}

\item{upsert_at_end_of_each_plan}{Do you want to upsert your results automatically at the end of each plan?}

\item{insert_at_end_of_each_plan}{Do you want to insert your results automatically at the end of each plan?}

\item{action_fn_name}{The name of the function that will be called for each analysis with arguments \code{data}, \code{argset}, \code{schema}}

\item{data_selector_fn_name}{The name of a function that will be called to obtain the data for each analysis. The function must have the arguments \code{argset}, \code{schema} and must return a named list.}

\item{schema}{A named list that maps \code{sc::config$schemas} for use in \code{action_fn_name} and \code{data_selector_fn_name}}

\item{info}{Information for documentation}
}
\description{
This function is used to easily create a task
}
