% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8_add_task_from_config.r
\name{task_from_config_v8}
\alias{task_from_config_v8}
\title{Creating a task from config (v8)}
\usage{
task_from_config_v8(
  name = NULL,
  name_grouping = NULL,
  name_action = NULL,
  name_variant = NULL,
  cores = 1,
  plan_argset_fn_name = NULL,
  for_each_plan = NULL,
  for_each_argset = NULL,
  universal_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action_fn_name,
  data_selector_fn_name = NULL,
  schema = NULL,
  info = NULL
)
}
\description{
This function is used to easily create a task
}
