% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_config_datetime.R
\name{update_config_datetime}
\alias{update_config_datetime}
\title{update_config_datetime
Updates the config_datetime db tables}
\usage{
update_config_datetime(type, tag, date = NULL, datetime = NULL)
}
\arguments{
\item{type}{a}

\item{tag}{a}
}
\description{
update_config_datetime
Updates the config_datetime db tables
}
