% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wiki_sc}
\alias{wiki_sc}
\title{Example simplicial complex from Wikimedia Commons}
\format{A bipartite \link[igraph:igraph]{igraph::igraph} object prefixed with class 'sc'.}
\source{
\url{https://commons.wikimedia.org/wiki/File:Simplicial_complex_example.png}
}
\usage{
wiki_sc
}
\description{
A 3-dimensional simplicial complex used for illustration.
}
\keyword{datasets}
