% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingmethods_pt1.R
\name{metropolis}
\alias{metropolis}
\title{Metropolis Algorithm (Supports Optional Metropolis-Hastings)}
\usage{
metropolis(target, initial, proposal, n = 1e+05, hastings = FALSE)
}
\arguments{
\item{target}{A function representing the target distribution (unnormalized density).}

\item{initial}{A function that generates the initial sample.}

\item{proposal}{A function that proposes a new sample given the current sample.}

\item{n}{Integer, number of iterations for the algorithm. Default is 100000.}

\item{hastings}{Logical, whether to use Metropolis-Hastings. Currently, only Metropolis is implemented. Default is FALSE.}
}
\value{
A list containing the sampled values.
}
\description{
This function implements the Metropolis algorithm for sampling from a given target distribution.
Currently, Metropolis-Hastings is not implemented, but the function structure allows future extension.
}
\examples{
# Define target distribution (unnormalized density)
target <- function(x) { dnorm(x, mean = 0, sd = 1) }

# Define proposal function (Normal random walk)
proposal <- function(x) { rnorm(1, mean = x, sd = 1) }

# Define initial value function
initial <- function() { 0 }

# Run Metropolis algorithm
samples <- metropolis(target, initial, proposal, n = 10000, hastings = FALSE)

# Plot the histogram of samples
hist(unlist(samples), probability = TRUE, main = "Metropolis Samples",
     xlab = "x", col = "lightblue", border = "black")

}
