% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplices.r
\name{simplices}
\alias{simplices}
\alias{vertices}
\title{Vertices and simplices af a simplicial complex object}
\usage{
vertices(sc, simplices = NULL)

simplices(sc, vertices = NULL)
}
\arguments{
\item{sc}{Bipartite \link[igraph:igraph]{igraph::igraph} object, representing a simplicial
complex.}

\item{vertices, simplices}{Numeric vertex IDs or character names of nodes of
\code{sc} whose incident nodes are desired (if any). Defaults to
\code{NULL}.}
}
\description{
These functions return the vertices and the simplices of an
object of class 'sc'.
}
\details{
\strong{sc} encodes simplicial complexes as bipartite \link[igraph:igraph]{igraph::igraph} objects
prefixed with a new 'sc' class, whose nodes are partitioned into those
encoding the \emph{vertices} versus defining \emph{simplices} of the complex. The
simplices are taken to be these defining simplicies together with all of
their \emph{faces} (sub-simplices).

The \code{"type"} attribute takes the values \code{FALSE} on the vertex nodes and
\code{TRUE} on the simplex nodes. The functions \code{vertices()} and \code{simplices()}
return the nodes of the 'sc' object corresponding to the vertices and to the
simplices, respectively. To obtain the vertices in a specific set of
simplices or the simplices containing a specific set of vertices, pass the
indices of these simplices or vertices to the corresponding arguments.
}
