% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rundate.R
\name{update_rundate}
\alias{update_rundate}
\title{update_rundate
Updates the rundate db tables}
\usage{
update_rundate(task)
}
\arguments{
\item{task}{a}
}
\description{
update_rundate
Updates the rundate db tables
}
